feat(component-playground): isolate each sketch in a preview iframe
Render each sketch in its own document (a second Vite HTML entry, sketch.html)
embedded as an iframe keyed by sketch id. Isolates CSS, DOM/global side
effects, and viewport per sketch, and gives a clean remount on selection —
the substrate the controls panel and editor harness build on.
