+++ apps/component-playground/src/sketch.html
@@ -1,0 +1,13 @@
+<!doctype html>
+<html lang="en">
+	<head>
+		<meta charset="UTF-8" />
+		<meta name="viewport" content="width=device-width, initial-scale=1" />
+		<title>sketch</title>
+		<link rel="stylesheet" href="./sketch.css" />
+	</head>
+	<body>
+		<div id="root"></div>
+		<script type="module" src="./sketch-entry.tsx"></script>
+	</body>
+</html>
